Revert "Merge pull request #15 from fuseio/add-accounts-to-ethereum"
This reverts commit 1f5987c636d9161a93fffb1621f4fdd0737f61e2, reversing
changes made to 3116319f79fa247c5282962f4692618cfe9d5aac.
